MVVM is a cursed architecture.  * It reads data from Omsi and binds nicely  - Can't set the variables in the datagrid
I've been looking and I can't find a neat solution to this. The ViewModel gets updated from two places but which of those updates should be sent back to Omsi?
Possible solutions:
 * Some clever and elegant solution using ReactiveUI (I feel like it must be able to do it but I have no idea how)
 * Stupidly complicated PropertyChangeNotification source tracking to work out who updated it so that we know whether or not to update Omsi
 * Hook off of another event in the View for specifically when a cell is edited (such as CellEditEnded) and use that to trigger the update back to Omsi (probs easiest, but requires code-behind)

MVVM is just not a very good design pattern and as much as I want this to be exemplary code, I'm not sure if we can do that with pure MVVM.
I'm not touching this again.